move container stuff to a plugin

+++ src/teststack/commands/containers.py
@@ -2,7 +2,6 @@ import json
 import os.path
 
 import click
-import docker.errors
 import jinja2
 
 from teststack import cli
@@ -12,83 +11,65 @@ from teststack import cli
 @click.option('--no-tests', '-n', is_flag=True, help='Don\'t start the tests container')
 @click.pass_context
 def start(ctx, no_tests):
-    client = docker.from_env()
+    client = ctx.obj['client']
     for service, data in ctx.obj['services'].items():
         name = f'{ctx.obj["project_name"]}_{service}'
-        try:
-            container = client.containers.get(name)
-        except docker.errors.NotFound:
-            client.containers.run(
+        container = client.container_get(name)
+        if container is None:
+            client.run(
                 image=data['image'],
                 ports=data.get('ports', {}),
-                detach=True,
                 name=name,
-                command=service.get('command', None),
+                command=data.get('command', None),
                 environment=data.get('environment', {}),
+                mount_cwd=False,
             )
 
     if no_tests is True:
         return
 
     env = ctx.invoke(cli.get_command(ctx, 'env'), inside=True, no_export=True, quiet=True)
-    try:
-        image = client.images.get(ctx.obj['tag'])
-    except docker.errors.ImageNotFound:
-        image = client.images.get(ctx.invoke(build))
+    image = client.image_get(ctx.obj['tag'])
+    if image is None:
+        image = client.image_get(ctx.invoke(build))
 
     name = f'{ctx.obj["project_name"]}_tests'
-    try:
-        container = client.containers.get(name)
-        if container.image.id != image.id:
-            end_container(container)
-            raise docker.errors.NotFound(message='Old Image')
-    except docker.errors.NotFound:
-
+    current_image_id = client.container_get_current_image(name)
+    if current_image_id != image:
+        client.end_container(name)
+        current_image_id = None
+    if current_image_id is None:
         command = ctx.obj['tests'].get('command', None)
         if command is None:  # pragma: no branch
             command = "sh -c 'trap \"trap - TERM; kill -s TERM -- -$$\" TERM; tail -f /dev/null & wait'"
 
-        container = client.containers.run(
+        container = client.run(
             image=image,
             stream=True,
             name=name,
             user=0,
             environment=env,
             command=command,
-            detach=True,
-            volumes={
-                os.getcwd(): {
-                    'bind': image.attrs['Config']['WorkingDir'],
-                    'mode': 'rw',
-                },
-            },
         )
+
     return container
-
-
-def end_container(container):
-    container.stop()
-    container.wait()
-    container.remove(v=True)
 
 
 @cli.command()
 @click.pass_context
 def stop(ctx):
-    client = docker.from_env()
+    client = ctx.obj['client']
     project_name = ctx.obj["project_name"]
     for service, _ in ctx.obj['services'].items():
         name = f'{project_name}_{service}'
-        try:
-            container = client.containers.get(name)
-        except docker.errors.NotFound:
+        container = client.container_get(name)
+        if container is None:
             continue
-        end_container(container)
-    try:
-        container = client.containers.get(f'{project_name}_tests')
-    except docker.errors.NotFound:
+        client.end_container(container)
+    container = client.container_get(f'{project_name}_tests')
+    if container is None:
         return
-    end_container(container)
+    client.end_container(container)
 
 
 @cli.command()
@@ -106,7 +87,9 @@ def restart(ctx):  # pragma: no cover
     default='Dockerfile.j2',
     help='template to render with jinja',
 )
-@click.option('--dockerfile', '-f', type=click.Path(), default='Dockerfile', help='dockerfile to write too')
+@click.option(
+    '--dockerfile', '--file', '-f', type=click.Path(), default='Dockerfile', help='container build file to write to'
+)
 @click.pass_context
 def render(ctx, template_file, dockerfile):
     env = jinja2.Environment(
@@ -121,15 +104,6 @@ def render(ctx, template_file, dockerfile):
     )
 
     template_string = template_file.read()
-
-    if 'commit' in ctx.obj:
-        template_string = '\n'.join(
-            [
-                template_string,
-                f'RUN echo "app-git-hash: {ctx.obj["commit"]} >> /etc/docker-metadata"',
-                f'ENV APP_GIT_HASH={ctx.obj["commit"]}\n',
-            ]
-        )
 
     template = env.from_string(
         '\n'.join(
@@ -148,24 +122,19 @@ def render(ctx, template_file, dockerfile):
 @cli.command()
 @click.option('--rebuild', '-r', is_flag=True, help='ignore cache and rebuild the container fully')
 @click.option('--tag', '-t', default=None, help='Tag to label the build')
-@click.option('--dockerfile', '-f', type=click.Path(), default='Dockerfile', help='dockerfile to write too')
+@click.option(
+    '--dockerfile', '--file', '-f', type=click.Path(), default='Dockerfile', help='container build file to write too'
+)
 @click.pass_context
 def build(ctx, rebuild, tag, dockerfile):
     ctx.invoke(render, dockerfile=dockerfile)
-    client = docker.from_env()
+    client = ctx.obj['client']
 
     if tag is None:
         tag = ctx.obj['tag']
 
     click.echo(f'Build Image: {tag}')
-
-    for chunk in client.api.build(path='.', dockerfile=dockerfile, tag=tag, nocache=rebuild, rm=True):
-        for line in chunk.split(b'\r\n'):
-            if not line:
-                continue
-            data = json.loads(line)
-            if 'stream' in data:
-                click.echo(data['stream'], nl=False)
+    client.build(dockerfile, tag, rebuild)
 
     return tag
 
@@ -174,7 +143,7 @@ def build(ctx, rebuild, tag, dockerfile):
 @click.pass_context
 def exec(ctx):  # pragma: no cover
     container = ctx.invoke(start)
-    os.execvp('docker', ['docker', 'exec', '-ti', container.id, 'bash'])
+    os.execvp('docker', ['docker', 'exec', '-ti', container, 'bash'])
 
 
 @cli.command()
@@ -183,20 +152,7 @@ def exec(ctx):  # pragma: no cover
 @click.pass_context
 def run(ctx, step, posargs):
     container = ctx.invoke(start)
-
-    def run_command(command):
-        command = command.format(
-            posargs=' '.join(posargs),
-        )
-        click.echo(click.style(f'Run Command: {command}', fg='green'))
-        socket = container.exec_run(
-            cmd=command,
-            tty=True,
-            socket=True,
-        )
-
-        for line in socket.output:
-            click.echo(line, nl=False)
+    client = ctx.obj['client']
 
     steps = ctx.obj['tests'].get('steps', {})
     if step:
@@ -206,6 +162,8 @@ def run(ctx, step, posargs):
     for command in commands:
         if isinstance(command, list):
             for cmd in command:
-                run_command(cmd)
+                click.echo(click.style(f'Run Command: {command}', fg='green'))
+                client.run_command(container, cmd.format(posargs=' '.join(posargs)))
         else:
-            run_command(command)
+            click.echo(click.style(f'Run Command: {command}', fg='green'))
+            client.run_command(container, command.format(posargs=' '.join(posargs)))
